+++ src/source/geojson_worker_source.test.ts
@@ -1,25 +1,19 @@
 import {describe, beforeEach, afterEach, test, expect, vi} from 'vitest';
-import {createGeoJSONIndex, GeoJSONWorkerSource, type LoadGeoJSONParameters} from './geojson_worker_source';
+import {GEOJSON_TILE_LAYER_NAME} from '@maplibre/vt-pbf';
+import {GeoJSONWorkerSource, type LoadGeoJSONParameters} from './geojson_worker_source';
 import {StyleLayerIndex} from '../style/style_layer_index';
 import {OverscaledTileID} from '../tile/tile_id';
-import {type LayerSpecification} from '@maplibre/maplibre-gl-style-spec';
-import {type Actor, type IActor} from '../util/actor';
-import {type TileParameters, type WorkerTileParameters, type WorkerTileResult, type WorkerTileWithData} from './worker_source';
 import {setPerformance, sleep} from '../util/test/util';
 import {type FakeServer, fakeServer} from 'nise';
-import {GEOJSON_TILE_LAYER_NAME} from '@maplibre/vt-pbf';
 import {SubdivisionGranularitySetting} from '../render/subdivision_granularity_settings';
-import {type WorkerTile} from './worker_tile';
+
+import type {GeoJSONVT} from '@maplibre/geojson-vt';
+import type {Actor, IActor} from '../util/actor';
+import type {TileParameters, WorkerTileParameters, WorkerTileResult, WorkerTileWithData} from './worker_source';
+import type {LayerSpecification} from '@maplibre/maplibre-gl-style-spec';
+import type {WorkerTile} from './worker_tile';
 
 const actor = {send: () => {}} as any as Actor;
-
-beforeEach(() => {
-    setPerformance();
-});
-
-afterEach(() => {
-    vi.clearAllMocks();
-});
 
 describe('geojson tile worker source', () => {
     const actor: IActor = {sendAsync: () => Promise.resolve({})} as any as IActor;
@@ -218,6 +212,7 @@ describe('reloadTile', () => {
         let data = await source.reloadTile(tileParams as any as WorkerTileParameters) as WorkerTileWithData;
         expect('rawTileData' in data).toBeFalsy();
         data.rawTileData = firstData.rawTileData;
+        data.encoding = 'mvt';
         expect(data).toEqual(firstData);
 
         // also shouldn't call loadVectorData again
@@ -279,6 +274,16 @@ describe('reloadTile', () => {
 });
 
 describe('resourceTiming', () => {
+    let server: FakeServer;
+    beforeEach(() => {
+        setPerformance();
+        global.fetch = null;
+        server = fakeServer.create();
+    });
+    afterEach(() => {
+        server.restore();
+        vi.clearAllMocks();
+    });
 
     const layers = [
         {
@@ -296,6 +301,9 @@ describe('resourceTiming', () => {
     } as GeoJSON.GeoJSON;
 
     test('loadData - url', async () => {
+        server.respondWith(request => {
+            request.respond(200, {'Content-Type': 'application/json'}, JSON.stringify(geoJson));
+        });
         const exampleResourceTiming = {
             connectEnd: 473,
             connectStart: 473,
@@ -321,14 +329,18 @@ describe('resourceTiming', () => {
 
         const layerIndex = new StyleLayerIndex(layers);
         const source = new GeoJSONWorkerSource(actor, layerIndex, []);
-        source.loadGeoJSONFromUrl = () => Promise.resolve(geoJson);
 
-        const result = await source.loadData({source: 'testSource', request: {url: 'http://localhost/nonexistent', collectResourceTiming: true}} as LoadGeoJSONParameters);
+        const promise = source.loadData({source: 'testSource', request: {url: 'http://localhost/nonexistent', collectResourceTiming: true}} as LoadGeoJSONParameters);
+        server.respond();
+        const result = await promise;
 
         expect(result.resourceTiming.testSource).toEqual([exampleResourceTiming]);
     });
 
     test('loadData - url (resourceTiming fallback method)', async () => {
+        server.respondWith(request => {
+            request.respond(200, {'Content-Type': 'application/json'}, JSON.stringify(geoJson));
+        });
         const sampleMarks = [100, 350];
         const marks = {};
         const measures = {};
@@ -352,9 +364,10 @@ describe('resourceTiming', () => {
 
         const layerIndex = new StyleLayerIndex(layers);
         const source = new GeoJSONWorkerSource(actor, layerIndex, []);
-        source.loadGeoJSONFromUrl = () => Promise.resolve(geoJson);
 
-        const result = await source.loadData({source: 'testSource', request: {url: 'http://localhost/nonexistent', collectResourceTiming: true}} as LoadGeoJSONParameters);
+        const promise = source.loadData({source: 'testSource', request: {url: 'http://localhost/nonexistent', collectResourceTiming: true}} as LoadGeoJSONParameters);
+        server.respond();
+        const result = await promise;
 
         expect(result.resourceTiming.testSource).toEqual(
             [{'duration': 250, 'entryType': 'measure', 'name': 'http://localhost/nonexistent', 'startTime': 100}]
@@ -483,13 +496,6 @@ describe('loadData', () => {
         await removePromise;
     });
 
-    test('loadData with geojson creates an non-updateable source', async () => {
-        const worker = new GeoJSONWorkerSource(actor, layerIndex, []);
-
-        await worker.loadData({source: 'source1', data: geoJson} as LoadGeoJSONParameters);
-        await expect(worker.loadData({source: 'source1', dataDiff: {removeAll: true}} as LoadGeoJSONParameters)).rejects.toBeDefined();
-    });
-
     test('loadData with geojson creates an updateable source', async () => {
         const worker = new GeoJSONWorkerSource(actor, layerIndex, []);
 
@@ -510,22 +516,6 @@ describe('loadData', () => {
         const result = await load1Promise;
         expect(result.data).toStrictEqual(updateableGeoJson);
         await expect(worker.loadData({source: 'source1', dataDiff: {removeAll: true}} as LoadGeoJSONParameters)).resolves.toBeDefined();
-    });
-
-    test('loadData with geojson network call creates a non-updateable source', async () => {
-        const worker = new GeoJSONWorkerSource(actor, layerIndex, []);
-
-        server.respondWith(request => {
-            request.respond(200, {'Content-Type': 'application/json'}, JSON.stringify(geoJson));
-        });
-
-        const promise = worker.loadData({source: 'source1', request: {url: ''}} as LoadGeoJSONParameters);
-
-        server.respond();
-
-        await promise;
-
-        await expect(worker.loadData({source: 'source1', dataDiff: {removeAll: true}} as LoadGeoJSONParameters)).rejects.toBeDefined();
     });
 
     test('loadData with diff updates', async () => {
@@ -558,104 +548,13 @@ describe('loadData', () => {
     });
 
     test('loadData should process cluster change with no data', async () => {
-        const mockCreateGeoJSONIndex = vi.fn(createGeoJSONIndex);
-        const worker = new GeoJSONWorkerSource(actor, layerIndex, [], mockCreateGeoJSONIndex);
-
+        const mockGeoJSONIndex = {
+            updateClusterOptions: vi.fn()
+        } as any as GeoJSONVT;
+        const worker = new GeoJSONWorkerSource(actor, layerIndex, [], () => mockGeoJSONIndex);
         await worker.loadData({source: 'source1', data: updateableFeatureCollection, cluster: false} as LoadGeoJSONParameters);
-        expect(mockCreateGeoJSONIndex.mock.calls[0][1].cluster).toBe(false);
-        await expect(worker.loadData({cluster: true} as LoadGeoJSONParameters)).resolves.toBeDefined();
-        expect(mockCreateGeoJSONIndex.mock.calls[1][1].cluster).toBe(true);
-    });
-});
-
-describe('getData', () => {
-    let server: FakeServer;
-    beforeEach(() => {
-        global.fetch = null;
-        server = fakeServer.create();
-    });
-    afterEach(() => {
-        server.restore();
-    });
-
-    const layers = [
-        {
-            id: 'layer1',
-            source: 'source1',
-            type: 'symbol',
-        },
-        {
-            id: 'layer2',
-            source: 'source2',
-            type: 'symbol',
-        }
-    ] as LayerSpecification[];
-
-    const geoJson = {
-        'type': 'Feature',
-        'geometry': {
-            'type': 'Point',
-            'coordinates': [0, 0]
-        }
-    } as GeoJSON.GeoJSON;
-
-    const updateableGeoJson = {
-        type: 'Feature',
-        id: 'point',
-        geometry: {
-            type: 'Point',
-            coordinates: [0, 0],
-        },
-        properties: {},
-    } as GeoJSON.GeoJSON;
-
-    const layerIndex = new StyleLayerIndex(layers);
-
-    test('getData returns correct geojson when the source was loaded with geojson', async () => {
-        const worker = new GeoJSONWorkerSource(actor, layerIndex, []);
-
-        await worker.loadData({source: 'source1', data: geoJson} as LoadGeoJSONParameters);
-        await expect(worker.getData()).resolves.toStrictEqual(geoJson);
-    });
-
-    test('getData after a geojson network call returns actual loaded geojson', async () => {
-        const worker = new GeoJSONWorkerSource(actor, layerIndex, []);
-
-        server.respondWith(request => {
-            request.respond(200, {'Content-Type': 'application/json'}, JSON.stringify(updateableGeoJson));
-        });
-
-        const load1Promise = worker.loadData({source: 'source1', request: {url: ''}} as LoadGeoJSONParameters);
-        server.respond();
-
-        await load1Promise;
-        await expect(worker.getData()).resolves.toStrictEqual(updateableGeoJson);
-    });
-
-    test('getData after diff updates returns updated geojson', async () => {
-        const worker = new GeoJSONWorkerSource(actor, layerIndex, []);
-
-        await worker.loadData({source: 'source1', data: updateableGeoJson} as LoadGeoJSONParameters);
-        await expect(worker.loadData({source: 'source1', dataDiff: {
-            add: [{
-                type: 'Feature',
-                id: 'update_point',
-                geometry: {type: 'Point', coordinates: [0, 0]},
-                properties: {}
-            }]
-        }} as LoadGeoJSONParameters)).resolves.toBeDefined();
-
-        await expect(worker.getData()).resolves.toStrictEqual({
-            type: 'FeatureCollection',
-            features: [
-                {...updateableGeoJson},
-                {
-                    type: 'Feature',
-                    id: 'update_point',
-                    geometry: {type: 'Point', coordinates: [0, 0]},
-                    properties: {}
-                }
-            ]
-        });
+        expect(mockGeoJSONIndex.updateClusterOptions).not.toHaveBeenCalled();
+        await expect(worker.loadData({updateCluster: true} as LoadGeoJSONParameters)).resolves.toBeDefined();
+        expect(mockGeoJSONIndex.updateClusterOptions).toHaveBeenCalled();
     });
 });